Changes type of String::length to size_t

+++ include/String.h
@@ -6,9 +6,9 @@
 extern "C" {
 #endif
 
-String* String_newFrom(const char* data, int len);
+String* String_newFrom(const char* data, size_t len);
 char* String_data(String* self);
-int String_len(String* self);
+size_t String_len(String* self);
 void String_release(String* self);
 
 #ifdef __cplusplus
